2021/3/21 21:56

+++ api/src/main/java/org/leafbook/api/respAbs/topicPage/EntryAbs.java
@@ -7,7 +7,7 @@ public class EntryAbs {
     private Long entryId;
     private String entryName;
     private String entryDesc;
-    private Integer likeNumber;
+    private Long likeNumber;
     private String entryAvatar;
     private Integer isLiked;
 }
